+++ packages/icons/src/index.ts
@@ -1,5 +1,5 @@
+export { default as ChatoolLogoIcon } from "./ChatoolLogoIcon";
 export { default as CheckIcon } from "./CheckIcon";
 export { default as ChevronDownIcon } from "./ChevronDownIcon";
 export { default as ChevronRightIcon } from "./ChevronRightIcon";
-export { default as KarnamehLogoIcon } from "./KarnamehLogoIcon";
 export { default as SpinnerIcon } from "./SpinnerIcon";