+++ apps/playground/CHANGELOG.md
@@ -6,5 +6,5 @@
 
 - Updated dependencies [9748740]
 - Updated dependencies [8da047b]
-  - @karnameh/icons@0.1.0
-  - @karnameh/ui@0.1.0
+  - @chatool/icons@0.1.0
+  - @chatool/ui@0.1.0